Adding cancancan.

+++ app/controllers/likes_controller.rb
@@ -2,7 +2,9 @@ class LikesController < ApplicationController
   def create
     post = Post.find params[:post_id]
     like = Like.new post: post, user: current_user
-    if like.save
+    if !(can? :like, post)
+      redirect_to root_path, alert: 'access denied'
+    elsif like.save
       redirect_to root_path
     else
       redirect_to root_path
@@ -11,7 +13,11 @@ class LikesController < ApplicationController
 
   def destroy
     like = current_user.likes.find params[:id]
-    like.destroy
-    redirect_to root_path
+    if can? :destroy, like
+      like.destroy
+      redirect_to root_path, notice: 'Like removed!'
+    else
+      redirect_to root_path, alert: 'access denied!'
+    end
   end
 end
