+++ app/controllers/members_controller.rb
@@ -2,7 +2,9 @@ class MembersController < ApplicationController
   def create
     post = Post.find params[:post_id]
     member = Member.new post: post, user: current_user
-    if member.save
+    if !(can? :member, post)
+      redirect_to root_path, alert: 'access denied'
+    elsif member.save
       redirect_to root_path
     else
       redirect_to root_path
@@ -11,7 +13,11 @@ class MembersController < ApplicationController
 
   def destroy
     member = current_user.members.find params[:id]
-    member.destroy
-    redirect_to root_path
+    if can? :destroy, member
+      member.destroy
+      redirect_to root_path, notice: 'Like removed!'
+    else
+      redirect_to root_path, alert: 'access denied!'
+    end
   end
 end
